Migrate SetPendingMoveType to UI ducks
