+++ src/scenes/Moves/MoveType.jsx
@@ -4,7 +4,6 @@ import windowSize from 'react-window-size';
 import { bindActionCreators } from 'redux';
 import PropTypes from 'prop-types';
 
-import { setPendingMoveType } from './ducks';
 import BigButton from 'shared/BigButton';
 import trailerGray from 'shared/icon/trailer-gray.svg';
 import truckGray from 'shared/icon/truck-gray.svg';
@@ -13,6 +12,7 @@ import './MoveType.css';
 
 import { mobileSize } from 'shared/constants';
 import { withContext } from 'shared/AppContext';
+import { setPendingMoveType } from '../../shared/UI/ducks';
 
 class BigButtonGroup extends Component {
   constructor() {
@@ -185,7 +185,10 @@ MoveType.propTypes = {
 };
 
 function mapStateToProps(state) {
-  return state.moves;
+  return {
+    ...state.moves,
+    pendingMoveType: state.ui.pendingMoveType,
+  };
 }
 
 function mapDispatchToProps(dispatch) {
