+++ src/shared/UI/ducks.js
@@ -9,6 +9,12 @@ const initialState = {
 };
 
 const SET_CURRENT_SHIPMENT_ID = 'SET_CURRENT_SHIPMENT_ID';
+const SET_PENDING_MOVE_TYPE = 'SET_PENDING_MOVE_TYPE';
+
+// Action Creation
+export function setPendingMoveType(value) {
+  return { type: SET_PENDING_MOVE_TYPE, payload: value };
+}
 
 export default function uiReducer(state = initialState, action) {
   switch (action.type) {
@@ -31,6 +37,10 @@ export default function uiReducer(state = initialState, action) {
         ...state,
         currentShipmentID: action.shipmentID,
       };
+    case SET_PENDING_MOVE_TYPE:
+      return Object.assign({}, state, {
+        pendingMoveType: action.payload,
+      });
     default:
       return state;
   }
